FTR-92 FTR-10 Extracted ReservationItem into ReservationItem.kt
