+++ app/src/main/java/hr/foi/air/servicesync/ui/contents/CalendarContent.kt
@@ -1,5 +1,6 @@
 package hr.foi.air.servicesync.ui.contents
 
+import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.padding
@@ -31,6 +32,7 @@ import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.ui.Alignment
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
@@ -44,6 +46,7 @@ import com.example.compose.primaryLight
 import com.example.compose.surfaceContainerDark
 import com.example.compose.surfaceContainerLight
 import hr.foi.air.servicesync.R
+import hr.foi.air.servicesync.ui.components.ReservationItem
 import hr.foi.air.servicesync.ui.components.isDark
 import kotlinx.coroutines.delay
 import java.util.concurrent.TimeUnit
@@ -70,7 +73,7 @@ fun CalendarContent(
     Column(
         modifier = modifier
             .fillMaxSize()
-            .padding(16.dp),
+            .padding(16.dp, 0.dp),
         verticalArrangement = Arrangement.spacedBy(8.dp),
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
@@ -103,7 +106,7 @@ fun CalendarContent(
                 text = stringResource(R.string.future_appointments),
                 style = MaterialTheme.typography.headlineSmall,
                 color = isDark(primaryDark, primaryLight),
-                modifier = Modifier.padding(bottom = 8.dp)
+                modifier = Modifier.padding(bottom = 1.dp).background(Color.Transparent),
             )
             LazyColumn(
                 modifier = Modifier.fillMaxSize(),
@@ -131,87 +134,3 @@ fun CalendarContent(
     }
 }
 
-@Composable
-fun ReservationItem(
-    companyName: String,
-    serviceName: String,
-    reservationDate: Long,
-    onClick: () -> Unit
-) {
-    val daysUntilReservation = calculateDaysUntilReservation(reservationDate)
-
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(8.dp)
-            //.background(isDark(surfaceDark, surfaceLight))
-            .clickable { onClick() },
-        shape = MaterialTheme.shapes.medium,
-        colors = CardDefaults.cardColors(
-            containerColor = isDark(surfaceContainerDark, surfaceContainerLight)
-        ),
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween
-        ) {
-            Column(
-                modifier = Modifier.weight(1f)
-            ) {
-                Text(
-                    text = serviceName,
-                    style = MaterialTheme.typography.bodyLarge.copy(
-                        fontWeight = FontWeight.Bold
-                    ),
-                    color = isDark(primaryDark, primaryLight)
-                )
-                Spacer(modifier = Modifier.height(2.dp))
-                Text(
-                    text = formatDate(reservationDate),
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = isDark(onSurfaceDark, onSurfaceLight)
-                )
-            }
-            Row (
-                verticalAlignment = Alignment.CenterVertically
-            ){
-                Icon(
-                    painter = painterResource(id = R.drawable.baseline_access_time_filled_24),
-                    contentDescription = "Clock Icon",
-                    modifier = Modifier.size(24.dp),
-                    tint = isDark(primaryDark, primaryLight)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Column (
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ){
-                    Text(
-                        text = daysUntilReservation.toString(),
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = isDark(primaryDark, primaryLight)
-                    )
-                    Spacer(modifier = Modifier.height(4.dp))
-                    Text(
-                        text = stringResource(R.string.days),
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = isDark(primaryDark, primaryLight)
-                    )
-                }
-            }
-        }
-    }
-}
-
-fun formatDate(timestamp: Long): String {
-    val sdf = java.text.SimpleDateFormat("dd.MM.yyyy HH:mm", java.util.Locale.getDefault())
-    return sdf.format(java.util.Date(timestamp))
-}
-
-fun calculateDaysUntilReservation(reservationDate: Long): Long {
-    val currentDate = System.currentTimeMillis()
-    val differenceInMillis = reservationDate - currentDate
-    return TimeUnit.MILLISECONDS.toDays(differenceInMillis)
-}
